+++ app/src/main/java/bme/msc/cookbook/model/RecipeBase.java
@@ -9,7 +9,7 @@ public interface RecipeBase {
 
     String getTotalTime();
 
-    double getRating();
+    float getRating();
 
     String getIngredients();
