+++ app/src/main/java/bme/msc/cookbook/model/orm/FavouriteRecipe.java
@@ -23,7 +23,7 @@ public class FavouriteRecipe extends SugarRecord implements RecipeBase {
 
     @SerializedName("rating")
     @Expose
-    private double rating;
+    private float rating;
 
     @SerializedName("ingredients")
     @Expose
@@ -44,7 +44,7 @@ public class FavouriteRecipe extends SugarRecord implements RecipeBase {
     public FavouriteRecipe() {
     }
 
-    public FavouriteRecipe(String name, String imgUrl, String totalTime, double rating,
+    public FavouriteRecipe(String name, String imgUrl, String totalTime, float rating,
                   String ingredients, String directions, String category, Date lastDate) {
         this.name = name;
         this.imgUrl = imgUrl;
@@ -80,11 +80,11 @@ public class FavouriteRecipe extends SugarRecord implements RecipeBase {
         this.totalTime = totalTime;
     }
 
-    public double getRating() {
+    public float getRating() {
         return rating;
     }
 
-    public void setRating(double rating) {
+    public void setRating(float rating) {
         this.rating = rating;
     }
