+++ app/src/main/java/bme/msc/cookbook/model/orm/OwnRecipe.java
@@ -21,7 +21,7 @@ public class OwnRecipe extends SugarRecord implements RecipeBase {
 
     @SerializedName("rating")
     @Expose
-    private double rating;
+    private float rating;
 
     @SerializedName("ingredients")
     @Expose
@@ -38,7 +38,7 @@ public class OwnRecipe extends SugarRecord implements RecipeBase {
     public OwnRecipe() {
     }
 
-    public OwnRecipe(String name, String imgUrl, String totalTime, double rating,
+    public OwnRecipe(String name, String imgUrl, String totalTime, float rating,
                            String ingredients, String directions, String category) {
         this.name = name;
         this.imgUrl = imgUrl;
@@ -73,11 +73,11 @@ public class OwnRecipe extends SugarRecord implements RecipeBase {
         this.totalTime = totalTime;
     }
 
-    public double getRating() {
+    public float getRating() {
         return rating;
     }
 
-    public void setRating(double rating) {
+    public void setRating(float rating) {
         this.rating = rating;
     }
